% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_rt_plot.R
\name{annotate_facet}
\alias{annotate_facet}
\title{Add annotation to a single facet}
\usage{
annotate_facet(
  geom = geom_text,
  condition,
  aes,
  fixed = list(color = "red"),
  ...
)
}
\arguments{
\item{geom}{Geometry function for the annotation}

\item{condition}{Which facet to draw in}

\item{aes}{aesthetic mapping}

\item{fixed}{Fixed aesthetics to use}

\item{...}{Values to annotate with, passed to a dataframe}
}
\value{
List of annotation geometries
}
\description{
Usually adding an annotation will draw something to every facet.
This helper constructs a dataframe out of the annotation parameters such that
the annotation is drawn in only a single facet.
}
