% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_rt_plot2.R
\name{compute_percent_change}
\alias{compute_percent_change}
\title{Title}
\usage{
compute_percent_change(difference_draws)
}
\arguments{
\item{difference_draws}{Output of `get_condition_difference_draws()`}
}
\value{
Dataframe with information about whether the 95% CrI (mean qi) contains
0 and the associated label for each tune:conditoin
}
\description{
Helper to wrangle the percent change labels for the posterior rt plot.
}
