% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_lists.R
\name{count_lists}
\alias{count_lists}
\title{Cross tabulate list assignments}
\usage{
count_lists(
  data,
  participant = "participant",
  critical = "critical_cond",
  polar = "polar_cond"
)
}
\arguments{
\item{data}{Experimental data of trials for each participant}

\item{participant}{Column name containing the participant IDs, usually either
participant or subj_id}

\item{critical}{Column name containing the critical condition list numbers,
usually critical_cond}

\item{polar}{Column name containing the polar condition list numbers,
usually polar_cond}
}
\value{
List of dataframes
}
\description{
Given experimental data with participants assigned to different critical
and polar conditions, cross tabulate the number of participants assigned
to each list
}
