% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{filter_exp_data}
\alias{filter_exp_data}
\title{Filter lexical decision data}
\usage{
filter_exp_data(subj_data)
}
\arguments{
\item{subj_data}{Individual subject data frame}
}
\value{
Filtered dataframe with a subset of renamed columns
}
\description{
Filters experiment data to only a subset of important columns
}
\examples{
\dontrun{
map_dfr(subj_data, filter_exp_data)
}

}
