% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{flag_trial_outliers}
\alias{flag_trial_outliers}
\title{Apply initial rt/rd transformations}
\usage{
flag_trial_outliers(
  data,
  by = "participant",
  rt_threshold = 1500,
  rd_threshold = 40
)
}
\arguments{
\item{data}{Experiment data with multiple participants' raw data}

\item{by}{Column name for the participant IDs, usually participant or subj_id}

\item{rt_threshold}{Numeric, default 1500, upper bound of RTs. Check overall
distribution first.}

\item{rd_threshold}{Numeric, default 40, lower bound of RDs. Check overall
distribution first.}
}
\value{
Data with additional columns.
}
\description{
Flags trials as an outlier if they are 3*IQR below Q1/above Q3 or if the RT
is above a given threshold based on visualizing the overall distribution.
Trials where the RD is too fast is treated as a motor error.
}
