% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_rt_plot2.R
\name{get_condition_difference_draws}
\alias{get_condition_difference_draws}
\title{Title}
\usage{
get_condition_difference_draws(
  model,
  dg,
  ndraws = NULL,
  TUNE_ORDER = c("lshll", "lhsll", "hll", "hlh", "lhslh", "lshlh")
)
}
\arguments{
\item{model}{brmsfit object}

\item{dg}{output of `marginal_data_grid()`}

\item{ndraws}{Number of draws, set to a small number (~100) to test. Default
NULL (all draws)}

\item{TUNE_ORDER}{Ordering of tunes, standard for project analysis}
}
\value{
Dataframe of the difference in epred logRT measures. Basically
the residual RT for each tune, which is interpretable as percent change from
some RT baseline. Eg -1.8 = 1.8% speedup.
}
\description{
Title
}
