% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{join_metrics}
\alias{join_metrics}
\title{Join in external data to experiment data}
\usage{
join_metrics(data, metrics)
}
\arguments{
\item{data}{Experimental data}

\item{metrics}{Elexicon metrics, see `load_metrics`}
}
\value{
data with lexical metrics
}
\description{
Merges in lexical measures and computes frequency ratios for each scale
}
