% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_rt_plot.R
\name{make_draws_legend}
\alias{make_draws_legend}
\title{Title}
\usage{
make_draws_legend(
  FACET_CONDITION = "lower",
  RECT_TOP = 559,
  RECT_HEIGHT = 16,
  CI_HEIGHT = 5,
  CI_LEFT = 1.5
)
}
\arguments{
\item{FACET_CONDITION}{String, which facet to draw the legend in}

\item{RECT_TOP}{Top y value for the legend rectangle}

\item{RECT_HEIGHT}{Height of the legend rectangle}

\item{CI_HEIGHT}{Height of the credible interval bars}

\item{CI_LEFT}{Left edge of the credible interval bars}
}
\value{
A list of geoms for drawing a legend inside a facet
}
\description{
Title
}
