% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_latex_switch.R
\name{make_latex_switch}
\alias{make_latex_switch}
\title{Create a latex switch command}
\usage{
make_latex_switch(
  model_coef_df,
  macroname = "statvalue",
  fstring = "$(\\\\hat\\\\beta = {estimate}, CI=[{conf.low},{conf.high}])$",
  remove_parens = TRUE,
  add_found_boolean = TRUE
)
}
\arguments{
\item{model_coef_df}{Dataframe containing model coefficient information}

\item{macroname}{Name to use for the latex command}

\item{fstring}{Formatted glue string to use for the inline citation. Use
column names that appear from the output of `broom.mixed::tidy()`}

\item{remove_parens}{Whether to remove parentheses from coefficient names,
such as in `(Intercept)`}

\item{add_found_boolean}{Whether to define a new boolean value `found` for the
latex command. This should be set to TRUE for the first switch command created,
but all following commands should use FALSE. Alternatively, you can create
the boolean globally in your LaTeX preamble and just always use FALSE with
this function.}
}
\value{
A latex command with the given name as a character vector for each
line of the command. The vector is returned invisibly, but the entire command
is printed via `cat()` so it can be easily copy and pasted into a document
as needed.
}
\description{
Given a dataframe, create a latex switch command.
}
