% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_rt_plot.R
\name{make_percent_labels}
\alias{make_percent_labels}
\title{Get speedup/slowdown percentages}
\usage{
make_percent_labels(
  pred_draws,
  TUNE_ORDER = c("lshll", "lhsll", "hll", "hlh", "lhslh", "lshlh")
)
}
\arguments{
\item{pred_draws}{Output of `add_epred_draws`}

\item{TUNE_ORDER}{Tune order, only a parameter in the event i need to change
it later}
}
\value{
dataframe with summarized percentage values for each tune in each condition
}
\description{
Given predicted draws, get the mean for each condition and compute the overall
percent speedup/slowdown for each tune. So, a value of -2.1 would equal a
2.1% speedup
}
