% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_latex_switch.R
\name{model_to_latex_switch}
\alias{model_to_latex_switch}
\title{Convert model to latex switch command}
\usage{
model_to_latex_switch(
  model,
  macroname = "statvalue",
  fstring = "$(\\\\hat\\\\beta = {estimate}, CI=[{conf.low},{conf.high}])$",
  roundto = 2,
  remove_parens = TRUE,
  add_found_boolean = TRUE
)
}
\arguments{
\item{model}{`brmsfit` or `mermod` model object}

\item{macroname}{Name to use for the latex command}

\item{fstring}{Formatted glue string to use for the inline citation. Use
column names that appear from the output of `broom.mixed::tidy()`}

\item{roundto}{Number of digits to round the numeric results to, defaults
to 2 decimal places.}

\item{remove_parens}{Whether to remove parentheses from coefficient names,
such as in `(Intercept)`}

\item{add_found_boolean}{Whether to define a new boolean value `found` for the
latex command. This should be set to TRUE for the first switch command created,
but all following commands should use FALSE. Alternatively, you can create
the boolean globally in your LaTeX preamble and just always use FALSE with
this function.}
}
\value{
A latex command with the given name as a character vector for each
line of the command. The vector is returned invisibly, but the entire command
is printed via `cat()` so it can be easily copy and pasted into a document
as needed.
}
\description{
Given a mixed model (either `brmsfit` or `mermod` object), extract the fixed
effect coefficient estimates and format the results as an inline statistic.
The output of this is a latex command definition that returns the inline
statistic given a coefficient name. This should be written to a .tex document
and imported into your latex project.
}
\details{
Note that this is a wrapper that processes the model coefficients for you.
If you need more control over the model coefficients, or want to use only
a subset of the coefficients, you should wrangle that dataframe and then use
`make_latex_switch()`.
}
