% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_rt_plot.R
\name{plot_posterior_rt_draws}
\alias{plot_posterior_rt_draws}
\title{Plot posterior draws}
\usage{
plot_posterior_rt_draws(
  pred_draws,
  intercept_lines,
  ylim = c(470, 559),
  scale = 1.2,
  facet_labels = c(lower = "Hear tough, see impossible", higher =
    "Hear impossible, see tough"),
  ...
)
}
\arguments{
\item{pred_draws}{Posterior expectation draws (from add_epred_draws)}

\item{intercept_lines}{Dataframe of intercept values to draw as lines}

\item{ylim}{plot y limits, passed to coord_cartesian}

\item{scale}{Scaling factor for text, should match scale argument in `save_plot`}

\item{facet_labels}{Named character to label facets, names must equal the
condition names. Defaults to the 2 condition case.}

\item{...}{Not used}
}
\value{
plot
}
\description{
Plots the posterior expectation draws for the higher and lower conditions,
colored by which conditions have pd > .95.
}
