% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_rt_plot2.R
\name{plot_posterior_rt_draws2}
\alias{plot_posterior_rt_draws2}
\title{Title}
\usage{
plot_posterior_rt_draws2(
  condition_difference_draws,
  scale = 1.2,
  facet_labels = c(lower = "Hear tough, see impossible", higher =
    "Hear impossible, see tough"),
  ...
)
}
\arguments{
\item{condition_difference_draws}{Output of `get_condition_difference_draws()`}

\item{scale}{Scale factor for plot output, default 1.2}

\item{facet_labels}{Labels for the condition facets, set NULL to exclude.}

\item{...}{Not used}
}
\value{
a ggplot
}
\description{
Title
}
