% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_latex_switch.R
\name{process_model_coef_df}
\alias{process_model_coef_df}
\title{Process model coefficients}
\usage{
process_model_coef_df(model, roundto = 2, remove_parens = TRUE)
}
\arguments{
\item{model}{`brmsfit` or `mermod` model object}

\item{roundto}{Number of digits to round the numeric results to, defaults
to 2 decimal places.}

\item{remove_parens}{Whether to remove parentheses from coefficient names,
such as in `(Intercept)`}
}
\value{
Dataframe of model fixed effect coefficients using `broom.mixed::tidy()`
}
\description{
A helper to process model coefficients in a consistent manner for writing
latex switch commands.
}
