% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save plot helper}
\usage{
save_plot(p, filename, dpi = 1200, ...)
}
\arguments{
\item{p}{Plot}

\item{filename}{String, filename to use, no file extension}

\item{dpi}{integer, dpi to use for png file}

\item{...}{Additional names parameters to pass}
}
\value{
Invisibly returns the paths to the saved plots
}
\description{
Saves a given plot as an svg, pdf, and png file with given parameters.
Recommended to use `ggview` to settle on the appropriate width, height, and
scale, then pass the parameters to this function
}
\details{
Assumes a `Figures/` directory in the current working directory
}
