% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_ppcheck}
\alias{save_ppcheck}
\title{Save posterior predictive check plots}
\usage{
save_ppcheck(mdl, filename, show = FALSE, pp_options = list(), ...)
}
\arguments{
\item{mdl}{Model fit with `brm`}

\item{filename}{Filename to use for files}

\item{show}{Logical, whether to return the plot object}

\item{pp_options}{List of named arguments to pass to `brms::ppcheck` such as
type or group}

\item{...}{Options to pass to save_plot}
}
\value{
plot from pp_check if show is TRUE, otherwise nothing
}
\description{
Save posterior predictive check plots
}
