% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{wrangle_covariates}
\alias{wrangle_covariates}
\title{Apply centering and transformations to covariates}
\usage{
wrangle_covariates(filtered_data)
}
\arguments{
\item{filtered_data}{Data with outliers filtered out already}
}
\value{
filtered_data with modified columns
}
\description{
Given experiment data that has been filtered for outliers, center various
measures.
}
