% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save plot helper}
\usage{
save_plot(
  p,
  filename,
  rootdir = "Figures",
  dpi = 1200,
  use_subfolders = FALSE,
  ...
)
}
\arguments{
\item{p}{Plot}

\item{filename}{String, filename to use, no file extension}

\item{rootdir}{Directory to save files to, defaults to Figures. If this does
not already exist it will be created. You can set this to a subdirectory
so that when you set `use_subfolders` to TRUE it will create the correct
subfolders where you want them.}

\item{dpi}{integer, dpi to use for png file}

\item{use_subfolders}{Logical, defaults to FALSE, if TRUE will create
pdf, png, and svg subfolders in the Figures/ directory and save each figure
to the appropriate subdirectory. Useful when you have a LOT of figures.}

\item{...}{Additional names parameters to pass}
}
\value{
Invisibly returns the paths to the saved plots
}
\description{
Saves a given plot as an svg, pdf, and png file with given parameters.
Recommended to use `ggview` to settle on the appropriate width, height, and
scale, then pass the parameters to this function
}
\details{
Default behavrior is to create a `Figures/` directory in the current working
directory.
}
